+++ aurum/internal/commands/app_command.py
@@ -5,7 +5,7 @@ import typing
 from hikari.permissions import Permissions
 from hikari.undefined import UNDEFINED
 
-from aurum.includable import Includable
+from aurum.internal.includable import Includable
 
 if typing.TYPE_CHECKING:
     from collections.abc import Sequence
